2103 move shared code for generating snapshot download url into image download utils
